Last of the known runtime null pointer situations.
When user asks for a column name or table that doesn't actually exists a
meaningful error message is returned.

Thought I check this in first because I realized that making it
interactive error fixing is going to be a bit more thought about what we
actually want to accomplish, and for what cases.