Now capturing table names that come before column names.
QueryRequest object now keeps track of all columns using a new
ColumnName object that contains a table a user specified, the column
name itself, and the type of the column. This way once SchemaInferrer is
fixed, we can use it to make sure the table that the user says a column
comes from is not inferred, but _definitely_ joined with. This can also
help the user prevent ambiguities with what they're asking.

This can't really be implemented until multiple joins and group-by
support for joins is stable.

Finally, we will be able to use this new data to translate back to the
user their result to be not just "name" but "customer name" - with the
rule of thumb that if they specified it to us, then we should specify it
back to them. (Versus if we said all of the inferred table names it
would be verbose and confusing to the user.)